Change LockService to an interface
