+++ src/main/java/org/opensearch/jobscheduler/rest/action/RestGetLockAction.java
@@ -19,7 +19,7 @@ import org.opensearch.jobscheduler.transport.AcquireLockResponse;
 import org.opensearch.jobscheduler.transport.AcquireLockRequest;
 import org.opensearch.jobscheduler.utils.JobDetailsService;
 import org.opensearch.jobscheduler.spi.LockModel;
-import org.opensearch.jobscheduler.spi.utils.LockService;
+import org.opensearch.jobscheduler.utils.LockServiceImpl;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.rest.BytesRestResponse;
 import org.opensearch.rest.RestRequest;
@@ -46,9 +46,9 @@ import static org.opensearch.jobscheduler.spi.LockModel.GET_LOCK_ACTION;
 public class RestGetLockAction extends BaseRestHandler {
     private final Logger logger = LogManager.getLogger(RestGetLockAction.class);
 
-    public LockService lockService;
+    public LockServiceImpl lockService;
 
-    public RestGetLockAction(final LockService lockService) {
+    public RestGetLockAction(final LockServiceImpl lockService) {
         this.lockService = lockService;
     }
