+++ src/main/resources/mappings/opensearch_job_scheduler_lock.json
