+++ src/test/java/org/opensearch/jobscheduler/rest/action/RestReleaseLockActionTests.java
@@ -19,7 +19,7 @@ import org.mockito.Mockito;
 import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.jobscheduler.JobSchedulerPlugin;
 import org.opensearch.jobscheduler.spi.LockModel;
-import org.opensearch.jobscheduler.spi.utils.LockService;
+import org.opensearch.jobscheduler.utils.LockServiceImpl;
 import org.opensearch.rest.RestHandler;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.test.OpenSearchTestCase;
@@ -32,7 +32,7 @@ public class RestReleaseLockActionTests extends OpenSearchTestCase {
 
     private RestReleaseLockAction restReleaseLockAction;
 
-    private LockService lockService;
+    private LockServiceImpl lockService;
 
     private String releaseLockPath;
 
@@ -45,7 +45,7 @@ public class RestReleaseLockActionTests extends OpenSearchTestCase {
         super.setUp();
         this.clusterService = Mockito.mock(ClusterService.class, Mockito.RETURNS_DEEP_STUBS);
         this.client = Mockito.mock(Client.class);
-        this.lockService = new LockService(client, clusterService);
+        this.lockService = new LockServiceImpl(client, clusterService);
         restReleaseLockAction = new RestReleaseLockAction(this.lockService);
         this.releaseLockPath = String.format(Locale.ROOT, "%s/%s/{%s}", JobSchedulerPlugin.JS_BASE_URI, "_release_lock", LockModel.LOCK_ID);
