+++ src/test/java/org/opensearch/jobscheduler/utils/LockServiceIT.java
@@ -6,7 +6,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.jobscheduler.spi.utils;
+package org.opensearch.jobscheduler.utils;
 
 import org.junit.Before;
 import org.junit.Ignore;
@@ -19,6 +19,7 @@ import org.opensearch.jobscheduler.spi.JobExecutionContext;
 import org.opensearch.jobscheduler.spi.LockModel;
 import org.opensearch.jobscheduler.spi.ScheduledJobParameter;
 import org.opensearch.jobscheduler.spi.schedule.Schedule;
+import org.opensearch.jobscheduler.spi.utils.LockService;
 import org.opensearch.test.OpenSearchIntegTestCase;
 
 import java.io.IOException;
@@ -93,7 +94,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
     public void testSanity() throws Exception {
         String uniqSuffix = "_sanity";
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -131,7 +132,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String lockID = "sanity_test_lock";
         String uniqSuffix = "_sanity";
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -165,7 +166,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_second_acquire";
         String lockID = randomAlphaOfLengthBetween(6, 15);
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -194,7 +195,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_long_lock_id";
         String lockID = randomAlphaOfLengthBetween(513, 1000);
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockService lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -216,7 +217,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_lock_release+acquire";
         String lockID = randomAlphaOfLengthBetween(6, 15);
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockService lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -248,7 +249,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_lock_expire";
         String lockID = randomAlphaOfLengthBetween(6, 15);
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         // Set lock time in the past.
         lockService.setTime(Instant.now().minus(Duration.ofSeconds(LOCK_DURATION_SECONDS + LOCK_DURATION_SECONDS)));
         final JobExecutionContext context = new JobExecutionContext(
@@ -282,7 +283,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
 
     public void testDeleteLockWithOutIndexCreation() throws Exception {
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockService lockService = new LockServiceImpl(client(), this.clusterService);
         lockService.deleteLock("NonExistingLockId", ActionListener.wrap(deleted -> {
             assertTrue("Failed to delete lock.", deleted);
             latch.countDown();
@@ -292,7 +293,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
 
     public void testDeleteNonExistingLock() throws Exception {
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         lockService.createLockIndex(ActionListener.wrap(created -> {
             if (created) {
                 lockService.deleteLock("NonExistingLockId", ActionListener.wrap(deleted -> {
@@ -314,7 +315,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_multi_thread_create";
         String lockID = randomAlphaOfLengthBetween(6, 15);
         CountDownLatch latch = new CountDownLatch(1);
-        final LockService lockService = new LockService(client(), this.clusterService);
+        final LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -377,7 +378,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_multi_thread_acquire";
         String lockID = randomAlphaOfLengthBetween(6, 15);
         CountDownLatch latch = new CountDownLatch(1);
-        final LockService lockService = new LockService(client(), this.clusterService);
+        final LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
@@ -444,7 +445,7 @@ public class LockServiceIT extends OpenSearchIntegTestCase {
         String uniqSuffix = "_lock_renew";
         String lockID = randomAlphaOfLengthBetween(6, 15);
         CountDownLatch latch = new CountDownLatch(1);
-        LockService lockService = new LockService(client(), this.clusterService);
+        LockServiceImpl lockService = new LockServiceImpl(client(), this.clusterService);
         final JobExecutionContext context = new JobExecutionContext(
             Instant.now(),
             new JobDocVersion(0, 0, 0),
